Updated Search
Split search into two pages one for results and one for loading searches
