+++ .idea/misc.xml
@@ -6,6 +6,7 @@
         <entry key="..\:/Users/Epikg/Documents/MovieDisplayApplication/app/src/main/res/layout/bookmarked_repos.xml" value="0.14722222222222223" />
         <entry key="..\:/Users/Epikg/Documents/MovieDisplayApplication/app/src/main/res/layout/movie_detail.xml" value="0.14722222222222223" />
         <entry key="..\:/Users/Epikg/Documents/MovieDisplayApplication/app/src/main/res/layout/movie_list_item.xml" value="0.12487335359675786" />
+        <entry key="..\:/Users/Epikg/Documents/MovieDisplayApplication/app/src/main/res/layout/movie_similar_search.xml" value="0.12273550724637682" />
         <entry key="app/src/main/res/layout/activity_bookmarked_repos.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.1503623188405797" />
         <entry key="app/src/main/res/layout/activity_repo_detail.xml" value="0.278125" />
