+++ app/src/main/res/layout/movie_search.xml
@@ -28,34 +28,6 @@
         style="?attr/materialButtonStyle"
         android:text="@string/search" />
 
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_search_results"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-        <TextView
-            android:id="@+id/tv_search_error"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="16dp"
-            android:textAppearance="?attr/textAppearanceBody1"
-            android:textColor="?attr/colorError"
-            android:visibility="invisible"
-            android:text="@string/search_error" />
-
-        <com.google.android.material.progressindicator.CircularProgressIndicator
-            android:id="@+id/loading_indicator"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:indeterminate="true"
-            android:layout_marginTop="32dp"
-            android:layout_gravity="top|center_horizontal"
-            android:visibility="invisible" />
-
-    </FrameLayout>
 
 </LinearLayout>
